README.txt
For this homework assignment, we learned how to store data locally on the phone using ionic. The interface I made includes two buttons, one changes the user name to "bryce", and the other removes that username. While this app is simple it was a good start to learning how to store information locally on a device. Eventually we will have to create a database for storing larger sets of complex data. This assignment allowed us to create a prototype for a settings page that we will use later in the project.